requirements.txt
tqdm
matplotlib
pylint
pytorch-lightning==1.3.5
ray==2.7.1
opencv_python
yacs>=0.1.8
joblib
einops==0.3.0
kornia==0.4.1
pickle5==0.0.11
timm>=0.3.2
hydra-core
omegaconf
wandb
rich
natsort
loguru
h5py
albumentations
torchmetrics==0.6.0
numpy
ipdb
